requirements.txt
Flask
python-dotenv
google-genai
requests
flask-cors